Reinstate magic finders for friendly_id
By default, friendly_id v5 turns off the magic finder overrides that allow
`Model.find(slug)` to work. We can configure friendly_id to include this
behaviour.

It would be a good idea to stop relying on this behaviour and instead either
call `Model.friendly.find(slug)` (which is the new way friendly_id recommends
finding by slug), or `Model.find_by(slug: slug)`. Until the test suite is green
and the Rails 4 upgrade is complete however, that is going to be very tricky.