Be explicit about datetime format we want
The default precision when coercing a DateTime to a JSON-encoded string has
changed between Rails 3 and 4 (it now includes 3 fractional digits by default).
This updates the presenter code to maintain the previous behaviour of presenting
the public_timestamp without the extra precision.

I also considered updating the failing DocumentFilterPresenterTest test to cope
with the new precision, but it felt safer to maintain the existing presentation.